Remove HEAD request on the create_url function
This fixes #6 

For some reason, our HEAD request that we made to validate URLs (to see if they are not only semantically correct but also valid "live" URLs) was erroring out for some sites on Netlify. In #6, mega.nz Cloudflare protection was blocking requests from Netlify (but not our local network).

This PR removes that for the time being until we find a better solution.